move job processes around so they can be called in a library (and dispatched from lambda)
