+++ src/workers/job-processes.js
@@ -1,0 +1,122 @@
+import { r } from '../server/models'
+import { sleep, getNextJob, updateJob, log } from './lib'
+import { exportCampaign, uploadContacts, assignTexters, createInteractionSteps, sendMessages, handleIncomingMessageParts } from './jobs'
+import { setupUserNotificationObservers } from '../server/notifications'
+
+
+/* Two process models are supported in this file.
+   The main in both cases is to process jobs and send/receive messages
+   on separate loop(s) from the web server.
+   * job processing (e.g. contact loading) shouldn't delay text message processing
+
+   The two process models:
+   * Run the 'scripts' in dev-tools/Procfile.dev
+ */
+
+const jobMap = {
+  "export": exportCampaign,
+  "upload_contacts": uploadContacts,
+  "assign_texters": assignTexters,
+  "create_interaction_steps": createInteractionSteps,
+  // usually dispatched in three separate processes, not here
+  "send_messages": sendMessages
+}
+
+export async function processJobs() {
+  setupUserNotificationObservers()
+  // eslint-disable-next-line no-constant-condition
+  while (true) {
+    try {
+      await sleep(1000)
+      const job = await getNextJob()
+      if (job) {
+        await (jobMap[job.job_type])(job)
+        await r.table('job_request')
+          .get(job.id)
+          .delete()
+      }
+
+      var twoMinutesAgo = new Date(new Date() - 1000 * 60 * 2)
+      // delete jobs that are older than 2 minutes
+      // to clear out stuck jobs
+      await r.knex('job_request')
+        .where({ assigned: true })
+        .where('updated_at', '<', twoMinutesAgo)
+        .delete()
+    } catch (ex) {
+      log.error(ex)
+    }
+  }
+}
+
+const messageSenderCreator = (subQuery) => {
+  return async () => {
+    setupUserNotificationObservers()
+    // eslint-disable-next-line no-constant-condition
+    while (true) {
+      try {
+        await sleep(1100)
+        await sendMessages(subQuery)
+      } catch (ex) {
+        log.error(ex)
+      }
+    }
+  }
+}
+
+export const messageSender01 = messageSenderCreator(function(mQuery) {
+  return mQuery.where(r.knex.raw("(contact_number LIKE '%0' OR contact_number LIKE '%1')"))
+})
+
+export const messageSender234 = messageSenderCreator(function(mQuery) {
+  return mQuery.where(r.knex.raw("(contact_number LIKE '%2' OR contact_number LIKE '%3' or contact_number LIKE '%4')"))
+})
+
+export const messageSender56 = messageSenderCreator(function(mQuery) {
+  return mQuery.where(r.knex.raw("(contact_number LIKE '%5' OR contact_number LIKE '%6')"))
+})
+
+export const messageSender789 = messageSenderCreator(function(mQuery) {
+  return mQuery.where(r.knex.raw("(contact_number LIKE '%7' OR contact_number LIKE '%8' or contact_number LIKE '%9')"))
+})
+
+export async function handleIncomingMessages() {
+  setupUserNotificationObservers()
+  // eslint-disable-next-line no-constant-condition
+  while (true) {
+    try {
+      const countPendingMessagePart = await r.knex('pending_message_part')
+      .count('id AS total').then( total => {
+        let totalCount = 0
+        totalCount = total[0].total
+        return totalCount
+      })
+
+      await sleep(500)
+      if(countPendingMessagePart > 0) {
+        await handleIncomingMessageParts()
+      }
+    } catch (ex) {
+      log.error(ex)
+    }
+  }
+}
+
+
+const processMap = {
+  'processJobs': processJobs,
+  'messageSender01': messageSender01,
+  'messageSender234': messageSender234,
+  'messageSender56': messageSender56,
+  'messageSender789': messageSender789,
+  'handleIncomingMessages': handleIncomingMessages
+}
+
+export async function dispatchProcesses(event, dispatcher) {
+  const toDispatch = event.processes || processMap
+  for (let p in toDispatch) {
+    if (p in processMap) {
+      dispatcher(p)
+    }
+  }
+}
