+++ src/workers/message-sender-56.js
@@ -1,22 +1,3 @@
-import { r } from '../server/models'
-import { log } from '../lib'
-import { sleep } from './lib'
-import { sendMessages } from './jobs'
+import { messageSender56 } from './job-processes'
 
-async function sendMyMessages() {
-  return sendMessages(function(mQuery) {
-    return mQuery.where(r.knex.raw("(contact_number LIKE '%5' OR contact_number LIKE '%6')"))
-  })
-}
-
-(async () => {
-  // eslint-disable-next-line no-constant-condition
-  while (true) {
-    try {
-      await sleep(1100)
-      await sendMyMessages()
-    } catch (ex) {
-      log.error(ex)
-    }
-  }
-})()
+messageSender56()
